requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2023.5.7     
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
idna==3.4
pytz==2023.3
requests==2.30.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2